working on generators refactor
