--- conans/model/__init__.py
@@ -1,3 +1,0 @@
-from .conan_generator import GeneratorManager, Generator
-
-registered_generators = GeneratorManager()
